+ 게임 오브젝트 순찰 기능 구현 + 루아 스크립트에 오브젝트 명령 기능   (스킬, 순찰, 공격) + 맵 내 오브젝트의 타입 추가 및 충돌 영역 설정   그에 따른 루아 스크립트 데이터 일부 수정 + ObjectController.cpp l_command_to_Object 함수 추가 및 루아 바인딩

+++ Classes/Objects/Animate.h
@@ -25,7 +25,7 @@ private:
 	const char* ani_name;
 	unsigned int rep_start_index, rep_end_index;
 	std::vector<cocos2d::CCSpriteFrame*>anim_frame_list;
-	int start_time, end_time;
+	long start_time, end_time;
 	float frame_interval_time;
 	unsigned long ani_index;
 };
